pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "cryptography>=44.0.3",
    "dnspython>=2.7.0",
    "geoip2>=5.1.0",
    "httpx>=0.28.1",
    "ipaddress>=1.0.23",
    "networkx>=3.4.2",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pyopenssl>=25.0.0",
    "python-nmap>=0.7.1",
    "python-whois>=0.9.5",
    "pywappalyzer>=0.1.1",
    "requests-futures>=1.0.2",
    "requests>=2.32.3",
    "streamlit>=1.45.0",
    "tldextract>=5.3.0",
    "trafilatura>=2.0.0",
    "whois>=1.20240129.2",
    "openssl-python>=0.1.1",
]
